requirement.txt
altgraph==0.17
appdirs==1.4.3
arcade==2.3.15
asgiref==3.2.7
astroid==2.4.2
attrs==19.3.0
auto-py-to-exe==2.7.8
beautifulsoup4==4.9.3
bottle==0.12.18
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.0
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cx-Freeze==6.1
cycler==0.10.0
distlib==0.3.1
Django==2.2
django-bootstrap3==14.2.0
django-debug-toolbar==3.2
django-extensions==3.1.0
djangorestframework==3.12.2
docutils==0.16
docx==0.2.4
Eel==0.12.4
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
identify==1.4.25
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mysqlclient==2.0.2
nodeenv==1.4.0
numpy==1.18.4
packaging==20.4
panda3d==1.10.6
parawrap==1.0
pefile==2019.4.18
Pillow==6.2.1
pluggy==0.13.1
pre-commit==2.6.0
py==1.9.0
pycparser==2.20
pyfiglet==0.8.post1
pygame==1.9.6
pygame-camera==0.1.1
pyglet==1.5.5
Pygments==2.7.4
pyinstall==0.1.4
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pylint==2.6.0
pymunk==5.6.0
PyMySQL==0.9.3
PyOpenGL==3.1.4
pyparsing==2.4.7
pypiwin32==223
pyspellchecker==0.5.4
python-dateutil==2.8.1
pytiled-parser==0.9.3
pytz==2020.1
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.24.0
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
textwrap3==0.9.2
toml==0.10.1
tox==3.18.1
translate==3.5.0
urllib3==1.25.10
virtualenv==20.4.0
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
wxPython==4.1.0
zope.event==4.5.0
zope.interface==5.1.2
